api/requirements.txt
Flask==2.2.3
pandas==2.0.0
numpy==1.24.3
yfinance==0.2.14
python-dateutil==2.8.2
gunicorn==20.1.0
